Changed generate_random_fraction function, implemented random decimal.

+++ Library-Functions-Folder/Library-Functions-Folder-8/library-functions-program-8.h
@@ -57,7 +57,8 @@ int* generate_fraction_variable(int,
 
 void fraction_variable_stdout(int*);
 
-int* generate_random_fraction(int,int);
+int* generate_random_fraction(double,
+  double);
 
 int fraction_variable_numerator(int*);
